requirements.txt
fastapi
tortoise-orm
aiosqlite
passlib[bcrypt]
python-jose[cryptography]
uvicorn
python-multipart
email-validator
